refactor: split MultiVariantCropperModal into hook, dial, geometry
681 lines held crop geometry math, a standalone straighten-dial
component, all cropper state/handlers, and the modal JSX.

Split into:
- multiVariantCropper.geometry.ts - angle/bounding-box/centered-crop math
- StraightenDial.tsx - the dial component, previously file-private
- useMultiVariantCropper.ts - crop state, orientation, confirm flow
- MultiVariantCropperModal.tsx - JSX only (681 -> 288 lines)

Pure extraction. Hook call order is preserved exactly as it was in the
component, so render behavior is unchanged. Verified the modal's Dialog
JSX and the StraightenDial body are byte-identical to main.

Export name and path unchanged, so all 5 consumers are untouched.

Verified: tsc -p tsconfig.app.json stays at 0 errors (lm-client has no
test suite configured).
